requirements.txt
ultralytics
streamlit
google-generativeai
python-dotenv
opencv-python
gTTS